Make money in wallet a double

+++ src/main/java/g9/springframework/silkroad/models/Customer.java
@@ -18,7 +18,7 @@ import lombok.Setter;
 @Entity
 @NoArgsConstructor
 public class Customer extends User {
-  private int moneyInWallet;
+  private double moneyInWallet;
 
   @OneToOne(cascade = CascadeType.ALL)
   @JoinColumn(name = "cart_id")
